Update app.R
Added code to include a req() checkpoint before the app pulls the parameter suite data for the chosen met station. The req() checkpoint  checks the existence of the file path to the chosen met/parameter combination, which ensures that the app has had enough time to repopulate the set of available parameter suites with those from the chosen met.