Update app.R
Added "at  3m" to the laynames for Relative Humidity that correspond to a variable name of "rh." Data listed as such in EDI is recorded at 3m even though it is unspecified. An extra space is added between "at" and "3m" because there is already a "Relative humidity at 3m" layname entry to correspond to the "rh3m" variable name.